+++ kernel/sm/wu/wustatetable.cpp
@@ -381,7 +381,7 @@ int  OccupyStateTableFirstVacantRow(StateTable *t, int *rowId, int rngBegin, int
 #ifdef _MSC_VER
 #pragma warning( default : 4146 )
 #endif
-		*rowId=bitId+(i-(uint32_t*)t->mem)*32;
+		*rowId=bitId+(int)(i-(uint32_t*)t->mem)*32;
 		ValidateGuardMemory(t,*rowId,NULL,ST_ROW_GUARD);
 		success=1;
 	}
